CRUD Accessorie

+++ app/Repository/Eloquent/AccessoriesRepository.php
@@ -30,21 +30,13 @@ class AccessoriesRepository extends BaseRepository implements AccessoriesReposit
         }
     }
 
-    public function store(Request $request): Model
-    {
-        try {
-            $accessories = Accessories::firstOrNew(['name' => $request->name, 'unit' => $request->unit]);
-            $accessories->save();
-            return $accessories;
-        } catch (\Throwable $th) {
-            throw $th;
-        }
-    }
-
     public function delete(String $id)
     {
         try {
             $accessories = Accessories::find($id);
+            if ($accessories->transaction()->get()->sum('qty') > 0) {
+                return false;
+            }
             return $accessories->delete();
         } catch (\Throwable $th) {
             throw $th;
